Move environment loader into dedicated module
