+++ verifiers/utils/env_utils.py
@@ -1,12 +1,9 @@
 from __future__ import annotations
 
-import importlib
-import inspect
-import logging
 import re
 import time
 from dataclasses import dataclass
-from typing import Callable, Literal, Protocol
+from typing import Literal
 
 from openai import AsyncOpenAI, BadRequestError
 from openai.types.chat.chat_completion import ChatCompletion
@@ -17,9 +14,6 @@ from openai.types.completion_choice import CompletionChoice
 from openai.types.completion_usage import CompletionUsage
 
 from verifiers.types import MessageType, ModelResponse
-
-class Environment(Protocol):
-    ...
 
 
 @dataclass(slots=True)
@@ -58,95 +52,6 @@ _VLLM_MAX_TOKENS_RE = re.compile(
     r"This model's maximum context length is (?P<limit>\d+) tokens and your request has "
     r"(?P<prompt_tokens>\d+) input tokens \((?P=completion_tokens) > (?P=limit) - (?P=prompt_tokens)\)\."
 )
-
-
-def load_environment(env_id: str, **env_args) -> Environment:
-    logger = logging.getLogger("verifiers.utils.env_utils")
-    logger.info(f"Loading environment: {env_id}")
-
-    module_name = env_id.replace("-", "_")
-    try:
-        module = importlib.import_module(module_name)
-
-        if not hasattr(module, "load_environment"):
-            raise AttributeError(
-                f"Module '{module_name}' does not have a 'load_environment' function. "
-                f"This usually means there's a package name collision. Please either:\n"
-                f"1. Rename your environment (e.g. suffix with '-env')\n"
-                f"2. Remove unneeded files with the same name\n"
-                f"3. Check that you've installed the correct environment package"
-            )
-
-        from verifiers.envs.environment import Environment as EnvironmentImpl
-
-        env_load_func: Callable[..., Environment] = getattr(
-            module, "load_environment"
-        )
-        sig = inspect.signature(env_load_func)
-        defaults_info = []
-        for param_name, param in sig.parameters.items():
-            if param.default != inspect.Parameter.empty:
-                if isinstance(param.default, (dict, list)):
-                    defaults_info.append(f"{param_name}={param.default}")
-                elif isinstance(param.default, str):
-                    defaults_info.append(f"{param_name}='{param.default}'")
-                else:
-                    defaults_info.append(f"{param_name}={param.default}")
-            else:
-                defaults_info.append(f"{param_name}=<required>")
-
-        if defaults_info:
-            logger.debug(f"Environment defaults: {', '.join(defaults_info)}")
-
-        if env_args:
-            provided_params = set(env_args.keys())
-        else:
-            provided_params = set()
-
-        all_params = set(sig.parameters.keys())
-        default_params = all_params - provided_params
-
-        if provided_params:
-            provided_values = []
-            for param_name in provided_params:
-                provided_values.append(f"{param_name}={env_args[param_name]}")
-            logger.info(f"Using provided args: {', '.join(provided_values)}")
-
-        if default_params:
-            default_values = []
-            for param_name in default_params:
-                param = sig.parameters[param_name]
-                if param.default != inspect.Parameter.empty:
-                    if isinstance(param.default, str):
-                        default_values.append(f"{param_name}='{param.default}'")
-                    else:
-                        default_values.append(f"{param_name}={param.default}")
-            if default_values:
-                logger.info(f"Using default args: {', '.join(default_values)}")
-
-        env_instance = env_load_func(**env_args)
-
-        if not isinstance(env_instance, EnvironmentImpl):
-            raise TypeError(
-                f"Environment '{env_id}' returned {type(env_instance)} which is not a verifiers Environment"
-            )
-
-        logger.info(f"Successfully loaded environment '{env_id}'")
-
-        return env_instance
-
-    except ImportError as e:
-        logger.error(
-            f"Failed to import environment module {module_name} for env_id {env_id}: {str(e)}"
-        )
-        raise ValueError(
-            f"Could not import '{env_id}' environment. Ensure the package for the '{env_id}' environment is installed."
-        ) from e
-    except Exception as e:
-        logger.error(
-            f"Failed to load environment {env_id} with args {env_args}: {str(e)}"
-        )
-        raise RuntimeError(f"Failed to load environment '{env_id}': {str(e)}") from e
 
 
 def infer_provider_name(client: AsyncOpenAI) -> str:
